refactor(Statement): fix typo in Statement component props
- Correct prop name from 'orgin' to 'origin' in Statement.vue

Why it worked?

Statement was a special vue component that renders a 'meta component'. When it does not have a `origin` property but the real selected component has a `origin` property, the real selected component will receive the `origin` property from Statement's parent Block. Keeping `origin` property here is a lot cleaner.