Dev/devcontainer (#276)
* fix(nested-fragments): adjust fragment width calculation

This commit addresses an issue where the width of a fragment was not being calculated correctly when it was nested within another interaction.

The `getOffsetX` function in `useFragmentData.ts` has been updated to account for the layers of nested interactions on the origin participant. This is done by calculating the depth of the origin participant and using that to determine the correct offset.

A new `centerToCenter` method has been added to the `Anchor2` class to calculate the distance between the centers of two anchors. This is used in the `getOffsetX` function to get a more accurate distance between the left participant and the origin.

The `Interaction.tsx` component has been updated to pass the correct `origin` prop to the underlying `div` element.

The snapshots for the nested interaction tests have been updated to reflect the changes in the fragment width calculation.

* feat: add devcontainer.json and dependabot.yml for development environment setup

* fix: update Playwright command to use 'pnpm dev' and add .pnpm-store to .gitignore

* feat: update devcontainer configuration for Node.js environment and adjust Playwright install command